2.0 - (!, read the description) fix RivalRebels with shaders, ! - (DOESN'T WORK)
/*It's actually very difficult to do anything when half of your keyboard doesn't work.*/

As for Rival Rebels, I have no idea why the fixes don't work, I did everything the same as for all the other fixes, but for some reason they don't work exactly with this mod.

 Possible assumption:
a strange and very confusing structure of the mod, which may prevent the use of mixins for it, as far as has been tested, this mod is not affected by mixins at all. That is, they load, but do nothing (not applied)